refactor(engine): widen display columns from u16 to u32
Grapheme::col, RowMap::locate/char_at, ViewportState/ScrollPosition's
horizontal_offset, and the format/style column pipeline all tracked
display columns as u16, saturating at 65,535 — nothing past that column
was addressable or scrollable on an unwrapped line. Widen the whole
document-column domain to u32; terminal-cell values (mouse coords,
gutter/content width, wrap widths, on-screen render columns) stay u16,
with debug_assert-guarded narrowing at the domain boundary once a
column is confirmed on-screen.

+++ hume-editor/src/editor/mod.rs
@@ -343,7 +343,7 @@ pub(crate) struct EditorState {
     /// Reusable scratch buffer for format operations in visual-line movement.
     pub(super) motion_format_scratch: hume_engine::format::FormatScratch,
     /// Reusable sticky-column buffer for visual j/k movement.
-    pub(super) visual_move_target_cols: Vec<u16>,
+    pub(super) visual_move_target_cols: Vec<u32>,
     /// The last repeatable editing action, available for replay via `.`.
     pub(super) last_repeatable_action: Option<RepeatableAction>,
     /// Accumulating selection-recipe buffer for the *next* edit's dot-repeat.
